Adds additional Releasing steps
While following this process recently, we came across a couple of steps
that may be worth explicitly calling out.

This commit adds a step to update any dependencies, which may update a
corresponding version snapshot file. For a Ruby gem, this may update the
gem's version in `Gemfile.lock`.

This also adds a note that after committing your changes, you should
push that commit up. This is so the commit exists for the tag to
reference when the tag itself is pushed.