Add dynamic templates (#26)
* Add reusable templates for dynamic READMEs

This is part of our initiative into having a single place
for some of our README sections. This way, when we have to update
the copyright or the images, we can do it at a single place.

* Update README image

The guidance I got for which image URL to use for our READMEs
was to use this canonical URL moving forward.