Add remaining repos to be updated when this workflow is triggered
All of these repos have had the workflow and snippet added to them.

Unfortunately, for this workflow to work, GitHub seems to need the repo list to be defined explicitly, instead of using `github.repository` or something else more generic.

Using a matrix, we can iterate over the repo list in the steps section. Not sure if there is a better way to do this, though.