Add remaining repos to be updated when this workflow is triggered (#30)
* Add remaining repos to be updated when this workflow is triggered

All of these repos have had the workflow and snippet added to them.

Unfortunately, for this workflow to work, GitHub seems to need the repo list to be defined explicitly, instead of using `github.repository` or something else more generic.

Using a matrix, we can iterate over the repo list in the steps section. Not sure if there is a better way to do this, though.

* Update README and footer sections with template

Also, add instructions on how to use the template
when creating a new repo.

* Document additional step for dynamic README

Besides adding the workflow to the new repo being created,
the new repo path needs to be added to the repo list matrix in the
dispatch workflow. This way, GitHub will propagate the footer
updates to all repos in the list.