Update dynamic README workflow to open a Pull Request (#31)
* Update dynamic render workflow

Some repositories have a protected branch rule. That means
our workflow won't work because commiting directly to main
is not allowed. To prevent that, the workflow will now
open a PR for the changes.

* Update workflow example

We only want to run this workflow if the README has changed, not
on every change.

* Removing leading / from snippet

As @nickcharlton pointed out,

> the leading / implies to me the start of a filesystem,
which it presumably isn't.

I tested removing it and it worked, so it was probably
needed when I was testing things but it's not needed anymore.

* Use latest workflow version