Add all public repositories
As part of adding a dynamic SECURITY.md file to our open source
projects, we need to have those projects listed here. Some of those
projects are already configured, and others are missing.

This commit adds all open source projects we want to update, and also
sorts them alphabetically, to make it easier to update the remaining
repositories.